💄 move advanced mode toggle
